+++ resources/views/admin/section-asset/index.php
@@ -3,30 +3,29 @@
 declare(strict_types=1);
 
 /**
- * @see AssetController::actionIndex()
+ * @see SectionAssetController::actionIndex()
  *
  * @var View $this
- * @var Entry|Section $model
+ * @var Section $model
  * @var AssetArrayDataProvider $provider
  */
 
-use Hirtz\Cms\Models\Entry;
 use Hirtz\Cms\Models\Section;
-use Hirtz\Cms\Modules\Admin\Controllers\AssetController;
-use Hirtz\Cms\Modules\Admin\Widgets\Navs\AssetHeader;
-use Hirtz\Cms\Modules\Admin\Widgets\Navs\AssetSubmenu;
+use Hirtz\Cms\Modules\Admin\Controllers\SectionAssetController;
+use Hirtz\Cms\Modules\Admin\Widgets\Navs\SectionHeader;
+use Hirtz\Cms\Modules\Admin\Widgets\Navs\SectionSubmenu;
 use Hirtz\Media\Modules\Admin\Data\AssetArrayDataProvider;
 use Hirtz\Media\Modules\Admin\Widgets\Grids\AssetGridView;
 use Hirtz\Media\Modules\Admin\Widgets\Navs\AssetModelActionDropdown;
 use Hirtz\Skeleton\Web\View;
 use Hirtz\Skeleton\Widgets\Grids\GridContainer;
 
-echo AssetHeader::make()
+echo SectionHeader::make()
     ->model($model)
     ->content(AssetModelActionDropdown::make()
         ->provider($provider));
 
-echo AssetSubmenu::make()
+echo SectionSubmenu::make()
     ->model($model);
 
 echo GridContainer::make()